Implemented MenuOverlay Vue component
- The MenuOverlay component is effectively our navigation component.
This navigation component does not have a toggle mechanism on it's
own, this aspect of component is actually managed by our
to-be-implemented RootLayout component, which will decide wether or
not the navigation component is active. This gives us more freedom
to use this navigation component in different ways within Layout
components lower in the hierarchy while not disturbing the overall
flow of the navigation throughout the app.