Implemented IconLink Vue component
- The IconLink component is a simple wrapper around a HTML link
element that uses our icon font to display the link as an icon.
This current implementation only supports external navigation,
that is, it is not possible to use this link component to navigate
to other pages of the app itself, for that it is necessary to use
an router-link component.