Implemented Projects page components
- The Projects page is split between two components, one for the
title section of the app and another for the contents of the page.
This page is meant to be used in conjunction with the SplitLayout
component. In our Contents section we're going to retrieve information
for a project from our markdown modules using the projectName
route parameter. Once that information is retrieved, we display it.