Change routes paths and names
The application's routes paths and names have been translated to
portuguese, this is mainly because the this website is targeted
mostly on the brazillian public.

besides that, small changes were also made to the
ProjectsTitleSection, not only to accomodate the new route names but
also other details like positioning and other small fixes to the
styling in general were made.