Improve SplitLayout left-column transitions
We are no longer applying a transition to every single element that
is a child of the left-column's .contents element, as this can cause
a small overhead to the browser and thus make the transitions render
a bit slower. Instead we'll make it so that the child component that
will be rendered within that section of the layout is required to
define it's own transitions in it's scoped stylesheet, so that we do
not apply an unecessary transition to elements that simply don't need
one, this will also allow each child component to have a custom
transition.

It's not a huge improvement but it will make the transitions slightly
smoother.