Rewrite MenuOverlay `$router` change detection callback
The callback in the MenuOverlay component that is called whenever
the $route variable changes (that this, usually when the user is
navigating) has been rewritten to make sure that the component's
`currentRoute` variable is always updated when the route changes.

In the previous code, the callback would listen to the root
element's transitionEnd event, so that the `currentRoute`
variable wouldn't suddenly change in the midst of the menu
transitioning out. The problem with that approach, however, is that
in cases where the menu is being hidden using CSS (e.g. it's
`display` property is set to `none` by a parent component), the
root element's transitionEnd event will never trigger, and the
`currentRoute` would not get updated. In order to fix that, a
callback is set to be called right after all of the components css
transitions end regardless of wheter the transitionEnd event fired or
not, this is done using a setTimeout function.